refactor: OrderPriceInfo, AmountList 컴포넌트 생성
