+++ src/components/common/index.ts
@@ -1,6 +1,6 @@
 export { default as PageExplanation } from './PageExplanation/PageExplanation';
 export { default as ShippingFeeInfo } from './ShippingFeeInfo/ShippingFeeInfo';
-export { default as OrderPrice } from './OrderPrice/OrderPrice';
+export { default as OrderPriceInfo } from './OrderPriceInfo/OrderPriceInfo';
 export { default as CartItemInfo } from './CartItemInfo/CartItemInfo';
 export { default as BasicButton } from './BasicButton/BasicButton';
 export { default as BackButton } from './BackButton/BackButton';
