+++ src/components/shoppingCart/index.ts
@@ -1,3 +1,4 @@
+export { default as AmountsList } from './AmountsList/AmountsList';
 export { default as PurchaseProcessLayout } from '@components/layout/PurchaseProcessLayout/PurchaseProcessLayout';
 export { default as CartListDescription } from './CartListDescription/CartListDescription';
 export { default as CartList } from './CartList/CartList';
